Moved all methods to header files
