+++ src/modules/keyboardmanager/common/KeyboardManagerConstants.h
@@ -81,4 +81,7 @@ namespace KeyboardManagerConstants
     // Flags used for distinguishing key events sent by Keyboard Manager
     inline const ULONG_PTR KEYBOARDMANAGER_SINGLEKEY_FLAG = 0x11;
     inline const ULONG_PTR KEYBOARDMANAGER_SHORTCUT_FLAG = 0x101;
+
+    // Dummy key event used in between key up and down events to prevent certain global events from happening
+    inline const DWORD DUMMY_KEY = 0xFF;
 }
