+++ src/modules/keyboardmanager/dll/KeyboardEventHandlers.h
@@ -4,9 +4,18 @@
 
 namespace KeyboardEventHandlers
 {
-    // Function to set the value of a key event based on the arguments
-    void SetKeyEvent(LPINPUT keyEventArray, int index, DWORD inputType, WORD keyCode, DWORD flags, ULONG_PTR extraInfo);
-
     // Function to a handle a single key remap
     intptr_t HandleSingleKeyRemapEvent(LowlevelKeyboardEvent* data, KeyboardManagerState& keyboardManagerState) noexcept;
+
+    // Function to a change a key's behavior from toggle to modifier
+    intptr_t HandleSingleKeyToggleToModEvent(LowlevelKeyboardEvent* data, KeyboardManagerState& keyboardManagerState) noexcept;
+
+    // Function to a handle a shortcut remap
+    intptr_t HandleShortcutRemapEvent(LowlevelKeyboardEvent* data, std::map<Shortcut, RemapShortcut>& reMap, std::mutex& map_mutex) noexcept;
+
+    // Function to a handle an os-level shortcut remap
+    intptr_t HandleOSLevelShortcutRemapEvent(LowlevelKeyboardEvent* data, KeyboardManagerState& keyboardManagerState) noexcept;
+
+    // Function to a handle an app-specific shortcut remap
+    intptr_t HandleAppSpecificShortcutRemapEvent(LowlevelKeyboardEvent* data, KeyboardManagerState& keyboardManagerState) noexcept;
 };
